support for one-time configs

+++ scripts/acq-edit.sh
@@ -17,11 +17,29 @@ changed=$?
 
 
 if [ $changed  -ne 0 ] ; then 
-  read -p "Are you sure you want to upload your config to $1? (y/N) " -n1 -r
+  echo "Are you sure you want to upload your config to $1? (y/N) "
+  read -p "(y/N) " -r
   echo 
   if [[ $REPLY =~ ^[Yy]$ ]] 
   then
-    scp $dest $1:/rno-g/cfg/acq.cfg
+    echo "If you want this to be a persistent config, type the word 'persistent'"
+    echo "Otherwise this will be treated as a one-time config, for which you may enter an identifiying tag (e.g. calib_092922) or have one generated for you if left blank."
+    read -r
+    echo 
+    if [[ "$REPLY" == "persistent" ]] 
+    then 
+      echo "Copying persistent config" 
+      scp $dest $1:/rno-g/cfg/acq.cfg
+    elif [ -z $REPLY ] 
+    then 
+      outname=`date -Is`.cfg 
+      echo "Using name $outname"
+      scp $dest $1:/rno-g/cfg/acq.cfg.once/$outname 
+    else 
+      outname="$REPLY.cfg"
+      echo "Using name $outname"
+      scp $dest $1:/rno-g/cfg/acq.cfg.once/$outname
+    fi 
   fi
 else 
   echo "no change!"
